+++ lib/phoenix_swagger/schema.ex
@@ -36,7 +36,9 @@ defmodule PhoenixSwagger.Schema do
     :properties,
     :additionalProperties,
     :discriminator,
-    :example]
+    :example,
+    :'x-nullable'
+  ]
 
   @doc """
   Construct a new %Schema{} struct using the schema DSL.
@@ -133,6 +135,7 @@ defmodule PhoenixSwagger.Schema do
       }
     }
   end
+
   def array(name) when is_atom(name) do
     %Schema{
       type: :array,
@@ -177,7 +180,9 @@ defmodule PhoenixSwagger.Schema do
   end
   def property(model = %Schema{type: :object}, name, type = %Schema{}, description, opts) do
     {required?, opts} = Keyword.pop(opts, :required)
+    {nullable?, opts} = Keyword.pop(opts, :nullable)
     property_schema = struct!(type, [description: type.description || description] ++ opts)
+    property_schema = if nullable?, do: %{property_schema | :'x-nullable' => true}, else: property_schema
     properties = (model.properties || %{}) |> Map.put(name, property_schema)
     model = %{model | properties: properties}
     if required?, do: required(model, name), else: model
@@ -215,7 +220,7 @@ defmodule PhoenixSwagger.Schema do
       [do: {:__block__, _, exprs}] -> exprs
       [do: expr] -> [expr]
     end
-    
+
     body =
       exprs
       |> Enum.map(fn {name, line, args} -> {:property, line, [name | args]} end)
@@ -707,4 +712,21 @@ defmodule PhoenixSwagger.Schema do
     |> required(property_name)
   end
 
+  @doc """
+  Sets the `x-nullable` vendor extension property for the schema.
+
+    ## Example
+
+    iex> alias PhoenixSwagger.Schema
+    ...> %Schema{type: :string} |> Schema.nullable(true)
+    %Schema{
+      type: :string,
+      'x-nullable': true
+    }
+  """
+  @spec nullable(%Schema{}, maybe_boolean) :: %Schema{}
+        when maybe_boolean: true | false | nil
+  def nullable(model = %Schema{}, value) when value in [true, false, nil] do
+    %{model | "x-nullable": value}
+  end
 end